Fin_Project/requirements.txt
Django==5.0.4
# Python==3.12.0
djangorestframework==3.15.1
pillow==10.3.0